requirements.txt
aiosqlite==0.20.0
anyio==4.3.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ollama==0.2.0
pytypedstream==0.1.0
sniffio==1.3.1
typing_extensions==4.11.0
